Added support for custom types in callback interfaces (#1178)
We just needed to implement `type_rs()`.  Amazingly, we were able to get away
with not defining this filter until now.

The other issue was that the Kotlin callback interface code depends on
AtomicLong, but it doesn't import that type.  I guess we were able to
get away with this because callback interfaces were always paired with
objects, which do import AtomicLong.