Update to Askama 0.11
The new Askama version adds refererences in less places.  This required
a few changes:
 - `&dyn fmt::Display` args get change to `&str`
 - Enum literals (`Type::String`) need a `.borrow()` appended to them if
   they get passed to a filter function.
 - Method calls that then get filtered, (`something.type_()|type_name`)
   need a `.borrow()` appended to them.  In a some cases, this can also
   be fixed by having the template code use a field from the template
   struct directly rather than via a method call.
 - Arguments captured in matches also need a `.borrow()` if they get
   passed to a filter function.

Fixed some incorrect whitespace handling code that was only working
because of Askama bugs that were fixed in this version.

Updated the Rust versions to match the policy for Firefox 97.  Askama
requires a newer minimum version, because it uses the `bool_to_option`
feature.