Make binding generation threadsafe (#1188)
Removed the `Config` field from the `CodeOracle` structs.  This allows
the struct to be state-less, which makes the filter code threadsafe.

The main thing we were using the config for was custom types, since the
type name depeneded on a config value.  Changed that so we always use
the name from the UDL file as the type name, then create a type alias in
the helper code.  PR #1189 takes this idea a lot further, this is like a
mini version of that.

The other type that needed changes was external types.  But this was
much simpler, we just needed to move some code from the `CodeType`
struct to `CodeDeclaration` struct.

This change indirectly makes the python ext-types test fail for me. I'm
pretty sure this is because of #1183 I just commented it out for now, I
think we are going to have a nice solution for this once the external
bindings crate code is ready to merge.