Reworked how we render type definitions (#1189)
Removed `CodeType` code that renders the type definition code, including
class definitions, FFIConverters, initialization code, imports, etc. from Kotlin.
Moved that functionality to the template code.

The initial motivation for this is #1188.  By moving this functionality
out of `CodeType`, we no longer need the `CodeOracle` to store the
`Config`.  It can now be stateless and the filters module is thread-safe
again. I think there's a nice symmetry to this. `CodeType` mirrors
`Type` in that it has access to the names of things but not the
definitions.

This also removes a bunch of boilerplate and puts the imports and
include_once calls are placed next to the code that uses them.

There are a couple drawbacks.  We now need an extra match statement for
Type in the template code.  The add_import/include_once_check code is
slightly hacky.  But I think it's a win overall.

Even though this doesn't change the python code at all, it makes the
python ext-types test fail for me.  I'm pretty sure this is because of
issue 1183  I just commented it out for now, I think we are going to have a
nice solution for this once the external bindings crate code is ready to
merge.