Updated the FFI naming and checksums
Updated how we calculate FFI names and especially how we calculate the
checksum part for functions.  Now the checksum is only derived from the
function definition, not the entire component interface.  This is needed
so the proc-macros can generate the checksums.  The new system should
still prevent the possibility of the bindings using one function
definition when the scaffolding was build using a different one.

Added the concept of a uniffi scaffolding contract.  We can bump that
whenever we make changes to how scaffolding calls work.  The FFI names
depend on this value rather than the UniFFI version, which should
prevent linker errors when uniffi and uniffi-bindgen don't have the
exact same version.

Consolidated the code into one module.  I think it's easier to
understand what's going on if you can see it in one place.

Removed some unneeded Result<> returns.

Updated the CI Rust version to 1.59, which is the current Firefox
minimum version.