Make `uniffi` the one crate that consumers need to depend on
Make `uniffi` re-export the functionality from `uniffi_build`,
`uniffi_bindgen`, and `uniffi_macros`.  This means it can be consumers
only need to depend on uniffi.  This slightly simplifies the process,
but the real reason is because it changes the versioning system.

Updated `uniffi` to be version `1.0.0`.  The idea here is that this is
the version number that consumers use, and we only need to bump it if we
make some backward compatible change to how UDL files are processed.

Updated the tutorial docs to describe the new system.  Also made them
recommend the system for managing `uniffi-bindgen` that we're using in
mozilla: create your own uniffi-bindgen binary so that your Cargo.toml
file controls the version and you can keep it in sync with the `uniffi`
used in your source code.