Added some extra remarks & pictures to readme
When I followed this readme there were many discrepancies and a lot of things weren't clear. Nothing was building for me, so I had to import the Xcode project and copy every single setting individually and match and test stuff around. Xcode is also very graphical and this contains 0 pictures.