fix: FFI functions for async code returns an `Option<Box<RustFuture<T>>>`.
Why? Because all types going through FFI must implement the `FfiDefault` trait.
It is basically used to define a default value to return in case of an error.
Before this patch, we were returning a `Box`, so a pointer… that was never
freed. It's a data leak. After this patch, we return `None` from `FfiDefault`,
which leaks nothing and has the same ABI representation for Rust.