Use SipHasher13 explicitly
DefaultHasher is not guaranteed to stay the same hasher in future
versions of rust. `SipHasher13` is the current hasher behind DefaultHasher
as of rust 1.66. It is worth noting that rust used to use `SipHasher24`
for HashMap, although back then it was not exposed as `DefaultHasher`,
but this is evidence that it can change.