Reworking FfiConverter type to work better with proc-macros
Right now the UDL and proc-macro code generation are at odds with each
other in how they handle `FfiConverter`. The UDL-based code generates a
separate FfiConverter unit struct for each type we want to move across
the FFI.  This allows us to generate FfiConverters for remote types
without breaking the orphan rules.

However, it requires the code that generates scaffolding functions to
know the full path to the FfiConverter struct.  This wasn't an issue
with the UDL-based code, but it doesn't work with the proc-macro code.
With proc-macros, you can have types with `#[uniffi::export]` defined
outside of the crate root, and this breaks the system since we no longer
know the path to the FfiConverter.

We've had some attemps at fixing this (#1250, #1373).  I hope this one
sticks.

This basically splits the difference between the UDL system and the
proc-macro system:
  - FfiConverter is now implemented on the type being moved acress the
    FFI, rather than a separate type.
  - FfiConverter gets an generic parameter that's basically arbitrary.
    We use it to get around the orphan rules.
  - We define a single `UniFFITag` struct in the crate root.
  - We implement `FfiConverter<UniFFITag>` on each type we want to
    move across the FFI.
  - Since `FfiConverter<UniFFITag>` contains a local type, it gets
    around the orphan rules.
  - The scaffolding code can get the correct FfiConverter impl using
    `<T as ::uniffi::FfiConverter<crate::UniFFITag>>` and this
    works regerdless of if `T` was defined in the crate root or a
    submodule.