feat: Resolve the async waker-re-entering problem in Kotlin.
This patch designs a “broken” Rust `Future` that calls the waker twice: Either
immediately, or a few seconds later. The 2 scenarios are problematics. Either
the waker is called multiple times while the first calls are not finished, or
the waker is called successively multiple times. In both case, it's an error
that must be handled gracefully.

This patch addresses the case for Kotlin first.

In Kotlin, we are using a `ConcurrentHashMap` to store the waker's
environments. Initially, we were doing:

1. `get` the environment,
2. Poll the future,
3. If the future is ready, then when cleaning up, `remove` the environment.

This patch changes that to the following:

1. `remove` the environment immediately,
2. Poll the future,
3. If the future is _NOT_ ready, then `put` the environment again.

Because the `remove` operation is “atomic”/synchronized, we are sure that the
guard (`if (env == null)`) will handle multiple problematic calls to the waker.
It acts as a “sync barrier”.