Make external types work with interfaces
The issue here is that interfaces have fundamental differences compared
to enums/records when it comes to moving them across the FFI since.
Interfaces use `Arc<>` on the Rust side and `void *` on the FFI side.

Added way to flag that external types are interfaces, updated the code
to handle these correctly.