Added return handling to `FfiConverter`
Added the `FfiConverter::lower_return` method.  This is like `lower()`
but specialized for scaffolding returns.  This allows us to always use a
single function to handle scaffolding calls, rather than
`call_with_output` or `call_with_result` depending on if the return type
is a `Result<>` or not.

Having a single code-path for return handling simplifies the code
generation, especially the macros.  We no longer need to try to parse
`Result<>` type names.  This is especially useful for crates that
type-alias their `Result<>` types.

Updated the async code to work with the new system.  Now `RustFuture`
only needs 1 generic argument.

Updated `object_references.md` and removed example code that's no longer
valid.  Replaced it with higher-level descriptions of what's going on,
hopefully this will stay not get out of date as quickly.