Update module.md
I thought about suggesting that the code be changed to always name the module `module.modulemap`, but including some documentation for folks who are attempting to create an XCFramework solves the same issue. (It took me an embarrassingly long time to find out that _not_ having this file named `module.modulemap` was the cause for an otherwise silent failure, reporting the module was unable to be imported in Swift.