Refactoring callback interface scaffolding code
The main goal here is to make the logic live in regular, non-templated,
code.  This makes it easier to understand what's going on and hopefully
easier to change.

One issue with this approach is that we need 3 functions depending on
the method return/throws types.  I think it's worth it to have clearer
code and I also hope to merge these into 1 function.  We could use the
same move as in PR#1469: define a new `FfiConverter` method like
`callback_interface_method_return()` that handles the specifics then
have a default implementation, with a specialized implementation for
`Result<>` and `()`.