Started work on the new callback interface ABI
The goal here is to lower the overhead of callback interface calls,
specifically targeting things like logging calls which can be called
many times in a short timespan.

Since this is an ABI change, it's not backwards compatible and therefore
requires a breaking version bump before it can be the default.

The initial ABI change was fairly minor: the argument RustBuffer is now
passed by reference instead of by value.  This means the foreign code
doesn't need to use an FFI call to destroy it.  For Kotlin, and maybe
other languages, it also seems to be slightly faster to use a reference
because that's less bits that need to be pushed to/popped from the
stack.

This brought some speedups to python and kotlin, and a minor speedup to
Swift that only passes the noise thresholdin the no-args-void-return
case:

callbacks/python-callbacks-basic
                        time:   [37.277 µs 37.299 µs 37.320 µs]
                        change: [-10.692% -10.532% -10.379%] (p = 0.00 < 0.05)
                        Performance has improved.

callbacks/kotlin-callbacks-basic
                        time:   [15.533 µs 15.801 µs 16.069 µs]
                        change: [-35.654% -31.481% -27.245%] (p = 0.00 < 0.05)
                        Performance has improved.

callbacks/swift-callbacks-no-args-void-return
                        time:   [454.96 ns 455.64 ns 456.71 ns]
                        change: [-35.872% -35.792% -35.710%] (p = 0.00 < 0.05)
                        Performance has improved.