Implement RustArcPtrUnsafe for object FFI free function
This allows FFI free function to have a separate type from regular
methods. This is important for C# bindings generator, because regular
`RustArcPtr` is represented as native C# `SafeHandle`, while
`RustArcPtrUnsafe` is represented as a `void*`. This distinction is
required to be able to call free function from inside
`SafeHandle::ReleaseHandle`.