Add "return_type()" to Constructor
This makes Constructor more nominal with Function, i.e. the constructor
can be used as a function in macros:
```
{% macro(func) %}
    {{ func.return_type() }}
{% endmacro %}

{% call macro(constructor) %}
```

This is very useful for generating Go bindings, because in Go errors
are handled via multiple return values, instead of exceptions. That
means that codegen for function return signatures get quite
complicated, and being able to re-use the same code for constructors as
for regular functions makes it much easier to generate code.