Fixes after the callback interface speedup merge
After this merged into main the performance tests failed because of a
clap issue.  I'm really not sure why this was, I don't think we upgraded
clap since I started the branch.  In any case, this fixes the issue.

Also, the `--compiler-messages` flag was not being honored.  Updated the
code to use this by replacing `RunScriptMode` enum, with
`RunScriptOptions`.  I think this makes more sense since we want the
performance test mode to have options.  I also think this will be easier
to extend going forward.