Ensure bridging header function prototypes are valid
Compiling my swift project I am seeing:

```
Generated/mycrateFFI.h:75:45: error: this function declaration is not a prototype [-Werror,-Wstrict-prototypes]
```

for the `ffi_mycrate_uniffi_contract_version` and
`uniffi_mycrate_checksum_func_myfun` declarations.

This is because they take no arguments which is emitted as `()`, for C `(void)`
is required, C++ is happy with either.

The included fixture attempts to validate that the generated header is valid
according to a compiler at a highly pendantic level, interestingly this doesn't
actually cover `-Wstrict-prototypes` so that is enabled separately.

Add a specific case to `swift::arg_list_ffi_decl` to handle this by adding the
`void` when there are no arguments. The resulting diff is:

```diff
--- swift_bridging_header_compileFFI.h.old	2023-04-06 14:30:29.994435300 +0000
+++ swift_bridging_header_compileFFI.h.new	2023-04-06 14:31:14.047300820 +0000
@@ -54,6 +54,7 @@

 void uniffi_swift_bridging_header_compile_fn_func_no_arguments(
     RustCallStatus *_Nonnull out_status
+
 );

 RustBuffer ffi_swift_bridging_header_compile_rustbuffer_alloc(
@@ -73,9 +74,11 @@
 );

 uint32_t ffi_swift_bridging_header_compile_uniffi_contract_version(
+    void

 );

 uint16_t uniffi_swift_bridging_header_compile_checksum_func_no_arguments(
+    void

 );
```