Adding the `foreignexecutor` module
This is the first step in an overall async plan layed out in PR #1494.
This allows Rust calls to be scheduled using the foreign executor (AKA
task queue, event loop, etc.).  It's not yet documented in the manual, I
plan to do that when I implement async callbacks.

I hope to use this to simplify the `RustFuture` code. Issue 1507 makes
me think that we don't want the foreign bindings to handle ownership of
the future, it's very tricky to manage this across an FFI.  Instead,
let's have Rust manage the future and have the foreign bindings only
handle the scheduling.

Added support for `usize' ints.  This lets foreign bindings choose
between using an int or pointer to represent an opaque type.

Kotlin: moved the async imports to Types.rs, so that we can use
`add_import()` to handle them.  The ForeignExecutor.kt template wants to
import some of the same items.

Swift: Added support for initialization functions