Reworked the future handling
The new system for async functions is:

1. You make a call to a generated async function in the foreign bindings.
2. That function suspends itself, then makes a scaffolding call.  In addition to the normal
    arguments, it passes a `ForeignExecutor` and callback.
3. Rust uses the `ForeignExecutor` to schedules polls of the Future until it's ready.  Then
    invokes the callback.
4. The callback resumes the suspended async function from (2), closing the loop.

This means that Rust manages the polling and the lifecycle of the
Future/waker. This leads to a nice simplification of both the Rust and
Foreign sides of the FFI.

One complication here is the handling of unit returns over the FFI.
With sync functions, a unit return can simply translate into a void
return with the C-ABI.  However, a void callback parameter is not valid,
so we use a u8 as a placeholder value.

Split up the error handling code so that we can use it for both sync and
async functions.

Added `FFIType.canonical_name()`.  This is used by some of the languages
to define future callback handler types.

Re-enabled the futures tests.  Hopefully we can get the docker image
worker. Lowered the delay time for the tests to make them run faster.
Give the Kotlin tests a bit more tolerance for slowness.  I was finding
they failed sometimes on my machine.