Refactoring Config code
Replaced `MergeWith` and `ExternalBindingsConfig` with `BindingsConfig`.
The main motivation is that `BindingsConfig` is less generic than
`MergeWith`, which means it can be match the specifics of config loading
more closely.  I plan to use this in the next commit.  As a bonus, we
don't need a separate trait for external bindings.  Also we don't need
to implement `From<&ComponentInterface>`, which always seemed a bit odd.