Refactoring the callback interface return code
Added a `FfiConverter` method to handle returning values from callback
interface methods.  This removes the need for multiple
`invoke_callback_*` methods and fixes #1547.

Removed the swift code that wrote an string to the RustBuffer for flat
errors.  No other bindings do that and we just ignored it on the Rust
side.

Implemented the lift/lower/read/write for the unit struct.  We now call
these methods in the callback interface code.