Reworking "crate mode" as "library mode"
With crate mode, the user specified a crate name and UniFFI handled
building it.  With library mode the user needs to build the library
themselves, then pass the path to UniFFI.

This is slightly less convenient for the user -- although not that much,
since somewhere in their build system they are going to need to build
the library and copy it out.  However, it's more flexible since users can
build their library with whatever combination of feature flags, targets,
etc. they want, and if this affects the generated bindings UniFFI will
work with it.  Before, UniFFI had to guess the flags to pass to `cargo
build`, which wasn't ideal.

Made a few doc changes to indicate that library mode is the recommended
way to build and will probably be the default sometime in the future.
However, for now it's still not the default.

Fixed some newer fixtures to use `include_scaffolding!()`