Refactored unexpected callback interface handling
Added a method to `FfiConverter` for this.  The default implementation
panics.  If we see the `uniffi::callback_error` attribute, then we
generate a function that uses the `From<UnexpectedUniFFICallbackError>`
impl.  That `uniffi::callback_error` attribute is set by the template
code if it sees that an error is used in a callback interface.

This is needed for callback interface support for proc-macros.