Implemented proc-macro callbacks
I used a separate module in the `proc-macro` fixture, which worked
really well with `cargo expand` to see how the code was being generated.

The trickiest part here is that the order of the methods matter, since
it's index is the methods list is part of the API contract.  To deal
with that I updated the `uniffi_meta` code to order the metadata items
and ensured that the methods are processed in the correct order.