Reworking the exported function code
Moved the `FnSignature` type into its own module and added more
functionality to it.  I'm hoping to reuse this for callback interface
methods.

Cleaned up and simplified `export::scaffolding`.  I've been hoping to do
this for a while and now seemed like a good time.

This is mostly just a refactor, but a few things changed:
  - Scaffolding functions now use the argument names rather than `argN`.
    I found it easier to code this way and I don't think we can ever
    really support nameless arguments since we need to use something on
    the foreign bindings side.
 - The error handling now uses `Error::new` rather than
   `Error::new_spanned`.  I don't think this will hurt the diagnostics
   much, but if we wanted to we could go back to `new_spanned()`, we
   would just need to clone the `syn::FnArg`.