Remove `FfiType::canonical_name()` and `Type::canonical_name()`
These exist in `uniffi_bindgen::interface` but is only used internally
by kotlin and swift - they're not used by the rest of `uniffi_bingen`
or any other bindings.
It creates a complicated relationship between the bindings and the
core, so it's better the complications live purely inside the bindings
that need it.