Format Python names according to PEP8
All internal symbols now start with `_Uniffi` or `_uniffi` so that name
clashes are unlikely. Additionally, this means that they are no longer
included in typical autocompletion. This is quite a heavy-handed
approach because it requires all changes to the Python bindings to be
aware that names have to start with an underscore.

Fixes #1557.