feat: Remove the `F` generic parameter from `RustFuture`.
`RustFuture` has a `future` field. It was: `UnsafeCell<F>`, now it
becomes `RwLock<Option<Pin<Box<dyn Future<Output = T>>>>>`.

First, `RwLock` is probably safer to use than `UnsafeCell` here.

Second, `Option` is required for the next commits (being able to cancel
a `Future`).

Third, `Pin` is required to keep the multi-thread safety, and to
simplify the next commits.

Fourth, `Box<dyn Future<Output = T>>` erases the need for the `F`
generic parameter

The primary goal of this patch is to remove the `F` generic parameter.