feat: Correctly drop `Future` when an async function is dropped or cancelled.
An async function in the foreign language side can be cancelled. For
example, in Kotlin:

```
var job = launch { someAsyncFunction() }
job.cancel()
```

The problem is: The `Future` running inside `someAsyncFunction` was not
dropped/cancelled. It can easily create dead-lock if some locks are used
inside `someAsyncFunction` because they are never released.

This patch focuses on Kotlin only for the codegen part. However, let's
start by explaining the modifications on the Rust side.

The procedural macros generate a function with a C ABI, that's not
new. What is new is that now this function returns a pointer to
the `RustFuture`. That's why `RustFuture::into_raw` is now public.
Previously, it was just calling `future.wake()` and nothing was returned
from this C function. But to be able to cancel/drop the `RustFuture`, we
need a pointer to it, otherwise the implementation is a complete black-
box from the foreign language point of view: the `RustFuture` is
consumed by `RustFuture::wake` (then leaked, then fetched again by the
waker etc.) and that's it. OK, so now, it returns a pointer to the
`RustFuture`.

The procedural macros generates a _new_ function with a C ABI to drop
a `RustFuture`. What it does is:

```rust
::uniffi::rust_call(uniffi_call_status, || {
    let uniffi_rust_future = ::uniffi::RustFuture::<#return_ty, crate::UniFfiTag>::from_raw(
        uniffi_future_ptr,
    );
    Ok(uniffi_rust_future.cancel_or_drop_inner_future())
});
```

First off, `RustFuture::from_raw` is now also public. Second, now
it becomes obvious why the `F` generic parameter of `RustFuture` was
creating a problem! How can we build a `RustFuture` if we don't know the
type of the `Future` itself? This generic parameter has been removed in
a previous patch (686f827aa).

What does `RustFuture::cancel_or_drop_inner_future`? It increments
the `wake_counter` so that it ensures the `RustFuture` _cannot_ be
awaken anymore. It also takes the value inside `RustFuture.future`,
leaving a `None`, which also means that the `RustFuture` cannot be
awaken too. Double guards. By taking the `Future`, we can drop it, thus
releasing all the code inside the `Future`. Our problem is now solved
from the Rust point of view!

Now, let's move on the foreign language land. This patch focuses on
Kotlin, another patch may come for Swift and Python soon. A
previous patch (0df55b22c) has switched `Continuation` for
`CancellableContinuation`. We can now install a `CompletionHandler` that
will be called by `CancellableContinuation.invokeOnCancellation`, but
also when the `CancellableContinuation.resume` or `.resumeWithException`
have been called. In all cases, we are sure the `CompletionHandler` is called.

What this `CompletionHandler` does is the following:

* Call the `…_uniffi_drop` FFI function,
* “Free” `callbackHolder` by freeing its previous value (the
  `callback`), hopefully to indicate the GC that it can be collected. It
  may solve the problem mentioned in
  https://github.com/mozilla/uniffi-rs/pull/1677#issuecomment-1660663156
  as a bonus.

The `completionHandler` is also passed to the function callback
handler, so that it can be called after `continuation.resume` or
`continuation.resumeWithException`.

Note that `completionHandler` receives `null` if it's a
completed or resumed termination, or a `CancellationException` if the
continuation has been cancelled. That's why it's valid to see
`continuation.resumeWithException(exception); completionHandler(null)`;
we must not mix those exceptions.

The rest of the patch is extra code to `FfiFunction` on the Rust side,
to add “companion” functions, it's a simple way to add extra functions
related to one `FfiFunction`.