fix: Ensure that `uniffi_self` lives long enough.
Now, `RustFuture` constraints the `Future` with the `'static` lifetime
bound. This patch ensures that `uniffi_self` lives long enough, because
`rustc` was complaining about it. Nonetheless, a new problem arised:
`RustBuffer` cannot be sent inside a `RustFuture` because it doesn't
implement `Send`. And so far, it seems safe to implement `Send` for
`RustBuffer`, which is what this patch does.

And now, everything seems to work.