fix: If `completionHandler` throws an exception, do no `resumeWithException`.
This patch fixes a subtle problem. If we are in the
`continuation.resume` path, and then `completionHandler`
throws an exception (for any reason), we will end up in
`continuation.resumeWithException`. Thus, the `continuation` will be
resumed twice.

To address this problem, this patches moves `completionHandler(null)`
inside a `finally` block.