fix: Avoind NPE if `Future` is cancelled really quickly.
It exists a very rare case where it's possible to cancel a `Job`, and
thus a `CoroutineScope` before the FFI binding building and running
the `RustFuture` can be called, and thus, no `rustFuturePtr` can be
computed.

In this rare case, the `completionHandler` will run `rustFuturePtr!!`
and boom.

This patch catches this case by checking that the `rustFuturePtr` isn't
`null` before calling the `…_drop` function to drop the `RustFuture`.