Fix Kotlin callback handler reference handling
When we execute a future on Kotlin, we create a `jni.Callback` instance
that Rust calls when the Future is ready.  We need to hold a reference
to that object or else the JVM will finalize it and when Rust tries to
invoke the callback bad things will happen. In the unit tests this
causes them to stall because the Kotlin coroutine continutation is never
completed.

The `callbackHolder` variable was intended to hold a reference, but it
doesn't seem to be working (see [arg0d's analysis](https://github.com/mozilla/uniffi-rs/pull/1677#issuecomment-1660663156)).
I belive the issue is that variable isn't actually used anywhere so
Kotlin frees it before the coroutine has completed and the reason #1684
fixes the issue is that it uses the variable in the completion handler
(https://github.com/mozilla/uniffi-rs/pull/1684/files#diff-f564b605a6ba32d19bcb47128dc9c5c2f01169e2a41fa60ffaaf0b6eea13845cR34)

This commit changes things so instead of using a local variable, there's
a global set that stores all active callbacks.  It seems to fix the
issue for me when using arg0d's `repeat(1000) { ... }` test.