Futures shared resource test updates
Set the timeout in Rust rather than the foreign language.  I'm not sure
exactly why, but on Kotlin if we see the issue the test simply hangs
rather than exiting with an exception.  By setting the timeout in Rust,
we can at least print out an error before the hangs.

Added Python and Swift tests.