Updating the foreign executor API
Added parameters/return values to signal failure.  In particular, we
need a way to signal that the foreign executor has been cancelled/closed
and therefore any futures that depend on it should be halted.  This is
needed to fix #1669.

Updated the Rust code to check for executor failures.  We need to
release any references passed to the waker callback since that callback
will not run in that case.

Updated the foreign code to handle cancelled futures.

Reworked and updated unit tests to test the new system.

Bumped `UNIFFI_CONTRACT_VERSION` since this requires changing the
scaffolding API.  I think this is correct, although you could make the
argument that it's not needed since the async code is still fairly
experimental.

Updated the `uniffi-versioning.md` doc now that
`UNIFFI_CONTRACT_VERSION` lives in `uniffi_meta`.